+++ CHANGELOG.md
@@ -36,6 +36,7 @@ Usage:
 * Fix #326: Add default `/metrics` path for `prometheus.io/path` annotation
   (Sort of [redundant](https://prometheus.io/docs/prometheus/latest/configuration/configuration/#scrape_config), this makes it explicit)
 * Fix #339: Https should be considered default while converting tcp urls
+* Fix: Quarkus supports S2I builds both for JVM and native mode
 
 ### 1.0.0-rc-1 (2020-07-23)
 * Fix #252: Replace Quarkus Native Base image with ubi-minimal (same as in `Dockerfile.native`)